+++ frontend/src/App.tsx
@@ -33,6 +33,7 @@ import PrintersPage from './pages/PrintersPage'
 import OrderPaymentsPage from './pages/OrderPaymentsPage'
 import CashShiftsPage from './pages/CashShiftsPage'
 import OrderingPointsPage from './pages/OrderingPointsPage'
+import OrderTransfersPage from './pages/OrderTransfersPage'
 import QrLandingPage from './pages/QrLandingPage'
 
 const ReportsPage = lazy(() => import('./pages/ReportsPage'))
@@ -115,6 +116,7 @@ function App() {
               <Route path="/order-payments" element={<ProtectedRoute permission="orders.create"><OrderPaymentsPage /></ProtectedRoute>} />
               <Route path="/cash-shifts" element={<ProtectedRoute permission="sales.edit"><CashShiftsPage /></ProtectedRoute>} />
               <Route path="/ordering-points" element={<ProtectedRoute permission="tables.manage"><OrderingPointsPage /></ProtectedRoute>} />
+              <Route path="/order-transfers" element={<ProtectedRoute permission="orders.transfer"><OrderTransfersPage /></ProtectedRoute>} />
               <Route
                 path="/products/:id/recipe"
                 element={